requirements.txt
kubernetes==32.0.1 ; python_version >= "3.11" and python_version < "3.12"
langchain-anthropic==0.2.4 ; python_version >= "3.11" and python_version < "3.12"
langchain-azure-ai==0.1.2 ; python_version >= "3.11" and python_version < "3.12"
langchain-chroma==0.1.4 ; python_version >= "3.11" and python_version < "3.12"
langchain-cohere==0.3.5 ; python_version >= "3.11" and python_version < "3.12"
langchain-community==0.3.19 ; python_version >= "3.11" and python_version < "3.12"
langchain-core==0.3.41 ; python_version >= "3.11" and python_version < "3.12"
langchain-elasticsearch==0.3.2 ; python_version >= "3.11" and python_version < "3.12"
langchain-experimental==0.3.4 ; python_version >= "3.11" and python_version < "3.12"
langchain-google-genai==2.0.11 ; python_version >= "3.11" and python_version < "3.12"
langchain-huggingface==0.1.2 ; python_version >= "3.11" and python_version < "3.12"
langchain-milvus==0.1.7 ; python_version >= "3.11" and python_version < "3.12"
langchain-ollama==0.2.3 ; python_version >= "3.11" and python_version < "3.12"
langchain-openai==0.3.7 ; python_version >= "3.11" and python_version < "3.12"
langchain-teddynote==0.3.42 ; python_version >= "3.11" and python_version < "3.12"
langchain-text-splitters==0.3.6 ; python_version >= "3.11" and python_version < "3.12"
langchain==0.3.20 ; python_version >= "3.11" and python_version < "3.12"
langchainhub==0.1.21 ; python_version >= "3.11" and python_version < "3.12"
uvicorn[standard]==0.34.0 ; python_version >= "3.11" and python_version < "3.12"
openai==1.65.4 ; python_version >= "3.11" and python_version < "3.12"
fastapi==0.115.11 ; python_version >= "3.11" and python_version < "3.12"
jinja2==3.1.6 ; python_version >= "3.11" and python_version < "3.12"
langgraph-checkpoint==2.0.17 ; python_version >= "3.11" and python_version < "3.12"
langgraph-prebuilt==0.1.2 ; python_version >= "3.11" and python_version < "3.12"
langgraph-sdk==0.1.55 ; python_version >= "3.11" and python_version < "3.12"
langgraph==0.3.5 ; python_version >= "3.11" and python_version < "3.12"
pydantic-core==2.27.2 ; python_version >= "3.11" and python_version < "3.12"
pydantic-settings==2.8.1 ; python_version >= "3.11" and python_version < "3.12"
pydantic==2.10.6 ; python_version >= "3.11" and python_version < "3.12"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "3.12"
pymermaid>=0.3.0
tiktoken>=0.5.0
tavily-python>=0.2.2
duckduckgo-search>=4.2.0
faiss-cpu>=1.7.4 ; python_version >= "3.11" and python_version < "3.12"
python-docx>=0.8.11 ; python_version >= "3.11" and python_version < "3.12"
unstructured>=0.10.18 ; python_version >= "3.11" and python_version < "3.12"
pymupdf>=1.22.5 ; python_version >= "3.11" and python_version < "3.12"
fastapi==0.115.11 ; python_version >= "3.11" and python_version < "3.12"
uvicorn[standard]==0.34.0 ; python_version >= "3.11" and python_version < "3.12"
pydantic==2.10.6 ; python_version >= "3.11" and python_version < "3.12"
pymermaid>=0.3.0 ; python_version >= "3.11" and python_version < "3.12"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "3.12"
tiktoken>=0.5.0 ; python_version >= "3.11" and python_version < "3.12"